refactor(codegen): extract array-method handlers from call.rs (Rule 4)
